backend/requirements.txt
alembic==1.4.2
bcrypt==3.1.7
cffi==1.14.0
click==7.1.2
dataclasses
fastapi==0.59.0
FastAPI-SQLAlchemy==0.1.0
h11==0.9.0
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.7.0
psycopg2==2.8.5
pydantic==1.5.1
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
redis==3.5.3
six==1.15.0
SQLAlchemy==1.3.18
starlette==0.13.4
uvicorn==0.11.5
websockets==8.1
